+++ add-dist.sh
@@ -2,7 +2,7 @@
 set -e
 
 echo "> npm run dist"
-npm run dist
+yarn run dist
 
 echo "> git add dist"
 git add dist