+++ src/components/Ruler/Ruler_snapshot.spec.js
@@ -1,0 +1,18 @@
+/* eslint-disable */
+// This file was automatically generated
+import React from 'react';
+import { shallow } from 'enzyme';
+import Ruler from './Ruler';
+
+describe('Ruler Component', () => {
+  describe('snapshots', () => {
+    it('should match snapshot', () => {
+    
+      const wrapper1 = shallow(
+        <Ruler />
+      );
+      expect(wrapper1).toMatchSnapshot();
+
+    });
+  });
+});